Resume button added
Resume is included in the downloads directory. Buttons to the resume link were added in the header and footer with the 'file' icon. I still need to check if the link leads to a github page. I could make it lead to a separate webpage hosted on my site, or simply download. There is also a resume link in the about section. I don't necessarily want it there, or not in that format.