Replaced "services" with "projects"
The template originally had its list of items and each of them are called "posts" or "services" in different places. Everywhere that it said "services" I replaced with "projects", including the /img/services directory. I also deleted all the crap in the readme and just put a link to the actual portfolio page.