Editing content for presentation
The stuff on here is ready to be viewed now. The resume link still doesn't work. Three files, navbar.html, main.html, project.html, were added to start trying to get separate pages for project descriptions. That isn't done yet, but I need to merge this with master to get the content on the site.